requirement.txt
# Core
Django==1.10.2
psycopg2
Pillow==3.3.1

# Batteries
django-mptt==0.8.6
django-mptt-admin
django-ckeditor==5.1.1
django-redactor==0.2
django-wysiwyg-redactor
django-redactoreditor==1.2.10
django-tinymce==2.3.0
django-suit-redactor==0.0.4
#django-filer==1.2.5

# REST

# Deploy
uwsgi

# dbmailer

# Other

# image
easy-thumbnails==2.3

# SEO
django-meta==1.3.2
https://github.com/madmusha/django-chunks2/archive/master.zip
django-seo2==1.0.1